Conflicts:
	cheatingday/src/main/java/com/icia/cheatingday/common/dto/Page.java
	cheatingday/src/main/java/com/icia/cheatingday/freeboard/controller/FreeBoardController.java
	cheatingday/target/classes/com/icia/cheatingday/common/dto/Page$PageBuilder.class
	cheatingday/target/classes/com/icia/cheatingday/common/dto/Page.class
	cheatingday/target/classes/com/icia/cheatingday/freeboard/controller/FreeBoardController.class
	cheatingday/target/m2e-wtp/web-resources/META-INF/MANIFEST.MF
	cheatingday/target/m2e-wtp/web-resources/META-INF/maven/com.icia/cheatingday/pom.properties
	cheatingday/target/test-classes/com/icia/cheatingday/FreeBoardServiceTest.class

+++ cheatingday/src/main/java/com/icia/cheatingday/user/service/mvc/UserService.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.*;
 import org.springframework.security.crypto.password.*;
 import org.springframework.stereotype.*;
 
+import com.icia.cheatingday.authority.dao.*;
 import com.icia.cheatingday.common.dto.*;
 import com.icia.cheatingday.exception.*;
 import com.icia.cheatingday.user.dao.*;
@@ -24,20 +25,22 @@ public class UserService {
 	@Autowired
 	private PointDao pointDao;
 	@Autowired
+	private FavoriteDao favoriteDao;
+	@Autowired
 	private PasswordEncoder pwdEncoder;
 	@Autowired
 	private ModelMapper modelMapper;
+	@Autowired
+	private AuthorityDao authDao;
 
 	// 내정보 읽기
 	public UserDto.DtoForRead myPage(String uUsername) {
 		User user = userDao.findById(uUsername);
-		if(user==null)
-			throw new UserNotFoundException();
+		//if(user==null)
+		//	throw new UserNotFoundException();
 		UserDto.DtoForRead dto = modelMapper.map(user, UserDto.DtoForRead.class);
-		
 		DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy년 MM월 dd일");
-		dto.setJoinDateStr(user.getUJoinDate().format(dtf));
-		
+		dto.setUJoinDateStr(user.getUJoinDate().format(dtf));
 		LocalDate joinDate = user.getUJoinDate().toLocalDate();
 		LocalDate today = LocalDate.now();
 		long days = ChronoUnit.DAYS.between(joinDate, today);
@@ -47,6 +50,7 @@ public class UserService {
 	// 비밀번호 확인
 	public void checkPwd(String uPassword, String uUsername) {
 		User user = userDao.findById(uUsername);
+		System.out.println(user);
 		if(user==null)
 			throw new UserNotFoundException();
 		String encodedPassword = user.getUPassword();
@@ -54,7 +58,7 @@ public class UserService {
 		if(result==false)
 			throw new JobFailException("비밀번호가 틀렸습니다");
 	}
-	// 비밀번호 변경
+	// 비밀번호 변경	
 	public void changePwd(String uPassword, String newUPassword, String UUsername) {
 		User user = userDao.findById(UUsername);
 		if(user==null)
@@ -67,24 +71,35 @@ public class UserService {
 		else 
 			throw new JobFailException("잘못된 비밀번호입니다.");
 	}
-	// 포인트
-	public Page point(int pageno,String uUsername) {
-		int countOfBoard = pointDao.count(uUsername);
+	// 포인트 리스트 페이징
+	public Page list(int pageno) {
+		int countOfBoard = pointDao.count();
 		Page page = PagingUtil.getPage(pageno, countOfBoard);
 		int srn = page.getStartRowNum();
 		int ern = page.getEndRowNum();
 		List<Point> pointList = pointDao.findAll(srn, ern);
-		List<UserDto.DtoForList> dtoList = new ArrayList<>();
+		List<PointDto.DtoForList> dtoList = new ArrayList<>();
 		for(Point point:pointList) {
-			UserDto.DtoForList dto = modelMapper.map(point, UserDto.DtoForList.class);
+			PointDto.DtoForList dto = modelMapper.map(point, PointDto.DtoForList.class);
 			dto.setAccumulationDayStr(point.getAccumulationDay().format(DateTimeFormatter.ofPattern("yyyy년 MM일 dd일")));
 			dtoList.add(dto);
 		}
-		page.setUlist(dtoList);
+		page.setPlist(dtoList);
 		return page;
+	}
+	// 즐겨찾기
+	public int favorite(int sNum, String uUsername ) {
+		int num = favoriteDao.findFavoriteById(uUsername, sNum);
+		if(num != 0) {
+			return num;
+		}
+		return 0;
 	}
 	// 회원탈퇴
 	public void resign(String uUsername) {
+		if(userDao.findById(uUsername)==null)
+			throw new JobFailException("회원 탈퇴에 실패했습니다");
 		userDao.delete(uUsername);
+		authDao.delete(uUsername);
 	}
 }
